requirements.txt
langgraph>=0.2.39
pydantic>=2.7.0
openai>=1.40.0
python-dotenv>=1.0.1
